# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "allow people to login into their account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "url to login",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "existing_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_views.UserCredentialReqStruct"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/login_mobile": {
            "post": {
                "description": "allow people to login into their account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "url to login with mobile number",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "existing_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_views.UserMobileCredentialReqStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/login_status": {
            "get": {
                "description": "api used to validate user login session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Login status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/sign_up": {
            "post": {
                "description": "allow people to create new to user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "url to signup",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "new_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_views.UserCredentialReqStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
...
        "/verify_social_auth": {
            "post": {
                "description": "url to signup/login with social authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Verify Social Authentication",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "new_or_existing_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_views.SocialAuthReqStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "my_modules.ResponseFormat": {
            "type": "object",
            "required": [
                "data",
                "msg",
                "status"
            ],
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "user_views.SocialAuthReqStruct": {
            "type": "object",
            "required": [
                "idToken"
            ],
            "properties": {
                "idToken": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user_views.UserCredentialReqStruct": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user_views.UserMobileCredentialReqStruct": {
            "type": "object",
            "required": [
                "mobile",
                "password"
            ],
            "properties": {
                "mobile": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}